requirements.txt
Django==4.2.5
django-extensions==3.2.3
djangorestframework==3.14.0
dr-scaffold==2.1.2
drf-yasg==1.21.7
isort==5.12.0
python-decouple==3.8
pytest-django==4.5.2
django-filter==23.3
pytest-django==4.5.2
autopep8==2.0.4
